authorization to task

+++ v1/controller/task.php
@@ -17,6 +17,83 @@ try {
     exit();
 }
 
+//begin auth script
+
+if ((!isset($_SERVER['HTTP_AUTHORIZATION'])) || (strlen($_SERVER['HTTP_AUTHORIZATION']) < 1)) {
+    $response = new Response();
+    $response->setHttpStatusCode(401);
+    $response->setSuccess(false);
+    (!isset($_SERVER['HTTP_AUTHORIZATION']) ? $response->addMessage("Access Token is missing from the header!") : false);
+    (strlen($_SERVER['HTTP_AUTHORIZATION']) < 1 ? $response->addMessage("Access Token cannot be blank!") : false);
+    $response->send();
+    exit();
+}
+
+$accessToken = $_SERVER['HTTP_AUTHORIZATION'];
+
+try {
+    $query = $writeDb->prepare('SELECT userid, accestokenexpiry, useractive, loginattempts 
+                                        FROM tblsessions, tblusers
+                                        WHERE tblsessions.userid = tblusers.id AND accestoken =:accesToken');
+    $query->bindParam(':accesToken', $accessToken, PDO::PARAM_STR);
+    $query->execute();
+
+    $rowCount = $query->rowCount();
+
+    if ($rowCount === 0) {
+        $response = new Response();
+        $response->setHttpStatusCode(401);
+        $response->setSuccess(false);
+        $response->addMessage("Invalid access token provided!");
+        $response->send();
+        exit();
+    }
+
+    $row = $query->fetch(PDO::FETCH_ASSOC);
+
+    $returned_userid = $row['userid'];
+    $returned_useractive = $row['useractive'];
+    $returned_loginattempts = $row['loginattempts'];
+    $returned_accestokenexpiry = $row['accestokenexpiry'];
+
+
+    if ($returned_useractive !== 'Y') {
+        $response = new Response();
+        $response->setHttpStatusCode(401);
+        $response->setSuccess(false);
+        $response->addMessage("User account is not active!");
+        $response->send();
+        exit();
+    }
+
+    if ($returned_loginattempts >= 3) {
+        $response = new Response();
+        $response->setHttpStatusCode(401);
+        $response->setSuccess(false);
+        $response->addMessage("User account is currently locked out!");
+        $response->send();
+        exit();
+    }
+
+    if (strtotime($returned_accestokenexpiry) < time()) {
+        $response = new Response();
+        $response->setHttpStatusCode(401);
+        $response->setSuccess(false);
+        $response->addMessage("Token has expired - please log in again!");
+        $response->send();
+        exit();
+    }
+} catch (PDOException $exception) {
+    $response = new Response();
+    $response->setHttpStatusCode(500);
+    $response->setSuccess(false);
+    $response->addMessage("There was an issue authenticating - Please log in again!");
+    $response->send();
+    exit();
+}
+
+//end auth script
+
 if (array_key_exists("taskId", $_GET)) {
 
     $taskId = $_GET['taskId'];
@@ -33,8 +110,11 @@ if (array_key_exists("taskId", $_GET)) {
     if ($_SERVER['REQUEST_METHOD'] === 'GET') {
 
         try {
-            $query = $readDb->prepare('SELECT id, title, description, DATE_FORMAT(deadline, "%d/%m/%Y %H:%i") as deadline, completed  from tbltasks WHERE id =:taskId');
+            $query = $readDb->prepare('SELECT id, title, description,
+            DATE_FORMAT(deadline, "%d/%m/%Y %H:%i") as deadline,
+            completed from tbltasks WHERE id =:taskId AND userid =:userId');
             $query->bindParam(':taskId', $taskId, PDO::PARAM_INT);
+            $query->bindParam(':userId', $returned_userid, PDO::PARAM_INT);
             $query->execute();
 
             $rowCount = $query->rowCount();
@@ -45,7 +125,6 @@ if (array_key_exists("taskId", $_GET)) {
                 $response->setSuccess(false);
                 $response->addMessage("Task not Found!");
                 $response->send();
-
                 exit();
             }
 
@@ -64,7 +143,6 @@ if (array_key_exists("taskId", $_GET)) {
             $response->toCache(true);
             $response->setData($returnData);
             $response->send();
-
             exit();
 
         } catch (TaskException $taskException) {
@@ -89,8 +167,9 @@ if (array_key_exists("taskId", $_GET)) {
     } elseif ($_SERVER['REQUEST_METHOD'] === 'DELETE') {
 
         try {
-            $query = $writeDb->prepare('DELETE from tbltasks WHERE id = :taskId');
+            $query = $writeDb->prepare('DELETE from tbltasks WHERE id = :taskId AND userid =:userId');
             $query->bindParam(':taskId', $taskId, PDO::PARAM_INT);
+            $query->bindParam(':userId', $returned_userid, PDO::PARAM_INT);
             $query->execute();
 
             $rowCount = $query->rowCount();
@@ -140,7 +219,6 @@ if (array_key_exists("taskId", $_GET)) {
                 $response->setSuccess(false);
                 $response->addMessage("Request body is not VALID json");
                 $response->send();
-
                 exit();
             }
 
@@ -180,14 +258,14 @@ if (array_key_exists("taskId", $_GET)) {
                 $response->setSuccess(false);
                 $response->addMessage("No task field provided to change!");
                 $response->send();
-
                 exit();
             }
 
             $query = $writeDb->prepare('SELECT id, title, description,
                     DATE_FORMAT(deadline, "%d/%m/%Y %H:%i") as deadline,
-                     completed  from tbltasks WHERE id =:taskId');
+                    completed  from tbltasks WHERE id =:taskId AND userid =:userId');
             $query->bindParam(':taskId', $taskId, PDO::PARAM_INT);
+            $query->bindParam(':userId', $returned_userid, PDO::PARAM_INT);
             $query->execute();
 
             $rowCount = $query->rowCount();
@@ -206,7 +284,7 @@ if (array_key_exists("taskId", $_GET)) {
                 $task = new Task($row['id'], $row['title'], $row['description'], $row['deadline'], $row['completed']);
             }
 
-            $queryString = "UPDATE tbltasks SET " . $queryFields . " WHERE id = :taskId";
+            $queryString = "UPDATE tbltasks SET " . $queryFields . " WHERE id = :taskId AND userid =:userId";
             $query = $writeDb->prepare($queryString);
 
             if ($title_updated === true) {
@@ -233,7 +311,8 @@ if (array_key_exists("taskId", $_GET)) {
                 $query->bindParam(":completed", $up_completed, PDO::PARAM_STR);
             }
 
-            $query->bindParam("taskId", $taskId, PDO::PARAM_INT);
+            $query->bindParam(":taskId", $taskId, PDO::PARAM_INT);
+            $query->bindParam(":userID", $returned_userid, PDO::PARAM_INT);
             $query->execute();
 
             $rowCount = $query->rowCount();
@@ -250,8 +329,9 @@ if (array_key_exists("taskId", $_GET)) {
 
             $query = $writeDb->prepare('SELECT id, title, description,
                     DATE_FORMAT(deadline, "%d/%m/%Y %H:%i") as deadline,
-                     completed  from tbltasks WHERE id =:taskId');
+                     completed  from tbltasks WHERE id =:taskId AND userid =:userId');
             $query->bindParam(':taskId', $taskId, PDO::PARAM_INT);
+            $query->bindParam(':userId', $returned_userid, PDO::PARAM_INT);
             $query->execute();
 
             $rowCount = $query->rowCount();
@@ -332,8 +412,9 @@ if (array_key_exists("taskId", $_GET)) {
 
         try {
             $query = $readDb->prepare('SELECT id, title, description, DATE_FORMAT(deadline, "%d/%m/%Y %H:%i") as deadline,
-                                        completed  from tbltasks WHERE completed =:completed');
+                                        completed  from tbltasks WHERE completed =:completed AND userid =:userId');
             $query->bindParam(':completed', $completed, PDO::PARAM_INT);
+            $query->bindParam(':userId', $returned_userid, PDO::PARAM_INT);
             $query->execute();
 
             $rowCount = $query->rowCount();
@@ -344,7 +425,6 @@ if (array_key_exists("taskId", $_GET)) {
                 $response->setSuccess(false);
                 $response->addMessage("Task not Found!");
                 $response->send();
-
                 exit();
             }
 
@@ -363,18 +443,14 @@ if (array_key_exists("taskId", $_GET)) {
             $response->toCache(true);
             $response->setData($returnData);
             $response->send();
-
             exit();
-
         } catch (TaskException $taskException) {
             $response = new Response();
             $response->setHttpStatusCode(500);
             $response->setSuccess(false);
             $response->addMessage($taskException->getMessage());
             $response->send();
-
             exit();
-
         } catch (PDOException $exception) {
             error_log("Database query error - " . $exception, 0);
             $response = new Response();
@@ -384,14 +460,12 @@ if (array_key_exists("taskId", $_GET)) {
             $response->send();
             exit();
         }
-
     } else {
         $response = new Response();
         $response->setHttpStatusCode(405);
         $response->setSuccess(false);
         $response->addMessage("Request method not allowed!");
         $response->send();
-
         exit();
     }
 } elseif (array_key_exists("page", $_GET)) {
@@ -412,9 +486,9 @@ if (array_key_exists("taskId", $_GET)) {
     if ($_SERVER['REQUEST_METHOD'] === 'GET') {
 
         try {
-            $query = $readDb->prepare('SELECT count(id) as totalNoOfTasks from tbltasks');
+            $query = $readDb->prepare('SELECT count(id) as totalNoOfTasks from tbltasks WHERE userid =:userId');
+            $query->bindParam(':userId', $returned_userid, PDO::PARAM_INT);
             $query->execute();
-
 
             $row = $query->fetch(PDO::FETCH_ASSOC);
             $tasksCount = intval($row['totalNoOfTasks']);
@@ -431,14 +505,15 @@ if (array_key_exists("taskId", $_GET)) {
                 $response->setSuccess(false);
                 $response->addMessage("Page Not Found!");
                 $response->send();
-
                 exit();
             }
 
             $offset = ($page == 1 ? 0 : ($limitPerPage * ($page - 1)));
 
             $query = $readDb->prepare('SELECT id, title, description, DATE_FORMAT(deadline, "%d/%m/%Y %H:%i") as deadline,
-                                        completed  from tbltasks LIMIT :pglimit OFFSET :offset');
+                                        completed  from tbltasks WHERE userid =:userId
+                                        LIMIT :pglimit OFFSET :offset');
+            $query->bindParam(':userId', $returned_userid, PDO::PARAM_INT);
             $query->bindParam(':pglimit', $limitPerPage, PDO::PARAM_INT);
             $query->bindParam(':offset', $offset, PDO::PARAM_INT);
             $query->execute();
@@ -450,7 +525,6 @@ if (array_key_exists("taskId", $_GET)) {
 
                 $tasksArray[] = $task->returnTaskAsArray();
             }
-
 
             $returnData = [];
             $returnData['rows_returned'] = $rowCount;
@@ -466,18 +540,14 @@ if (array_key_exists("taskId", $_GET)) {
             $response->toCache(true);
             $response->setData($returnData);
             $response->send();
-
             exit();
-
         } catch (TaskException $taskException) {
             $response = new Response();
             $response->setHttpStatusCode(500);
             $response->setSuccess(false);
             $response->addMessage($taskException->getMessage());
             $response->send();
-
             exit();
-
         } catch (PDOException $exception) {
             error_log("Database query error - " . $exception, 0);
             $response = new Response();
@@ -487,14 +557,12 @@ if (array_key_exists("taskId", $_GET)) {
             $response->send();
             exit();
         }
-
     } else {
         $response = new Response();
         $response->setHttpStatusCode(405);
         $response->setSuccess(false);
         $response->addMessage("Request method not allowed!");
         $response->send();
-
         exit();
     }
 } elseif (empty($_GET)) {
@@ -502,7 +570,10 @@ if (array_key_exists("taskId", $_GET)) {
     if ($_SERVER['REQUEST_METHOD'] === 'GET') {
 
         try {
-            $query = $readDb->prepare('SELECT id, title, description, DATE_FORMAT(deadline, "%d/%m/%Y %H:%i") as deadline, completed  from tbltasks');
+            $query = $readDb->prepare('SELECT id, title, description,
+                                DATE_FORMAT(deadline, "%d/%m/%Y %H:%i") as deadline,
+                                completed  from tbltasks WHERE userid =:userId');
+            $query->bindParam(':userId', $returned_userid, PDO::PARAM_INT);
             $query->execute();
 
             $rowCount = $query->rowCount();
@@ -513,7 +584,6 @@ if (array_key_exists("taskId", $_GET)) {
                 $response->setSuccess(false);
                 $response->addMessage("Tasks not Found!");
                 $response->send();
-
                 exit();
             }
 
@@ -532,18 +602,14 @@ if (array_key_exists("taskId", $_GET)) {
             $response->toCache(true);
             $response->setData($returnData);
             $response->send();
-
             exit();
-
         } catch (TaskException $taskException) {
             $response = new Response();
             $response->setHttpStatusCode(500);
             $response->setSuccess(false);
             $response->addMessage($taskException->getMessage());
             $response->send();
-
             exit();
-
         } catch (PDOException $exception) {
             error_log("Database query error - " . $exception, 0);
             $response = new Response();
@@ -553,7 +619,6 @@ if (array_key_exists("taskId", $_GET)) {
             $response->send();
             exit();
         }
-
     } elseif ($_SERVER['REQUEST_METHOD'] === 'POST') {
 
         try {
@@ -574,7 +639,6 @@ if (array_key_exists("taskId", $_GET)) {
                 $response->setSuccess(false);
                 $response->addMessage("Request body is not VALID json");
                 $response->send();
-
                 exit();
             }
 
@@ -585,7 +649,6 @@ if (array_key_exists("taskId", $_GET)) {
                 !isset($jsonData->title) ? $response->addMessage("Title field is mandatory and must be provided!") : false;
                 !isset($jsonData->completed) ? $response->addMessage("Completed field is mandatory and must be provided!") : false;
                 $response->send();
-
                 exit();
             }
 
@@ -600,12 +663,13 @@ if (array_key_exists("taskId", $_GET)) {
             $completed = $newTask->getCompleted();
 
             $query = $writeDb->prepare("INSERT INTO tbltasks
-                    (title, description, deadline, completed)
-                    VALUES (:title, :description, STR_TO_DATE(:deadline, '%d/%m/%Y %H:%i'), :completed)");
+                    (title, description, deadline, completed, userid)
+                    VALUES (:title, :description, STR_TO_DATE(:deadline, '%d/%m/%Y %H:%i'), :completed, :userId)");
             $query->bindParam(':title', $title, PDO::PARAM_STR);
             $query->bindParam(':description', $description, PDO::PARAM_STR);
             $query->bindParam(':deadline', $deadline, PDO::PARAM_STR);
             $query->bindParam(':completed', $completed, PDO::PARAM_STR);
+            $query->bindParam(':userId', $returned_userid, PDO::PARAM_INT);
             $query->execute();
 
             $rowCount = $query->rowCount();
@@ -616,7 +680,6 @@ if (array_key_exists("taskId", $_GET)) {
                 $response->setSuccess(false);
                 $response->addMessage("Failed to create task!");
                 $response->send();
-
                 exit();
             }
 
@@ -624,8 +687,9 @@ if (array_key_exists("taskId", $_GET)) {
 
             $query = $writeDb->prepare('SELECT id, title, description,
                         DATE_FORMAT(deadline, "%d/%m/%Y %H:%i") as deadline,
-                        completed  from tbltasks WHERE id =:lastTaskId');
+                        completed  from tbltasks WHERE id =:lastTaskId AND userid = :userId');
             $query->bindParam(':lastTaskId', $lastTaskID, PDO::PARAM_INT);
+            $query->bindParam(':userId', $returned_userid, PDO::PARAM_INT);
             $query->execute();
 
             $rowCount = $query->rowCount();
@@ -636,7 +700,6 @@ if (array_key_exists("taskId", $_GET)) {
                 $response->setSuccess(false);
                 $response->addMessage("Failed to retrieve task after creation!");
                 $response->send();
-
                 exit();
             }
 
@@ -655,18 +718,14 @@ if (array_key_exists("taskId", $_GET)) {
             $response->addMessage("Task created successfully");
             $response->setData($returnData);
             $response->send();
-
             exit();
-
         } catch (TaskException $taskException) {
             $response = new Response();
             $response->setHttpStatusCode(500);
             $response->setSuccess(false);
             $response->addMessage($taskException->getMessage());
             $response->send();
-
             exit();
-
         } catch (PDOException $exception) {
             error_log("Database query error - " . $exception, 0);
             $response = new Response();
@@ -674,7 +733,6 @@ if (array_key_exists("taskId", $_GET)) {
             $response->setSuccess(false);
             $response->addMessage("Failed to insert task into database - check submitted data for errors!");
             $response->send();
-
             exit();
         }
     } else {
@@ -683,16 +741,13 @@ if (array_key_exists("taskId", $_GET)) {
         $response->setSuccess(false);
         $response->addMessage("Request method not allowed!");
         $response->send();
-
         exit();
     }
-
 } else {
     $response = new Response();
     $response->setHttpStatusCode(404);
     $response->setSuccess(false);
     $response->addMessage("Endpoint not found!");
     $response->send();
-
     exit();
 }
